RMQ/requirements.txt
click==8.1.2
pandas==1.4.2
pika==1.2.0
psycopg2-binary==2.9.3
Flask==2.1.2
dataclasses=0.6
PyMySQL=1.0.2